more dependecy checking to _backend
